+++ src/patterns.js
@@ -73,7 +73,6 @@ import pat_url from './core/url.js';
 import pat_validation from './pat/validation/validation.js';
 import pat_zoom from './pat/zoom/zoom.js';
 import pat_push_kit from './core/push_kit.js';
-// import pat_chosen from './pat/chosen/chosen.js';
 */
 
 import pat_fullscreen from "./pat/fullscreen/fullscreen.js";